% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simData}
\alias{simData}
\title{Simulated data with multivariate binary responses}
\format{
An object of class \code{list} of length 8.
}
\usage{
simData
}
\description{
Simulated data set for a quick test. The data set is a list with 8
components: binary responses \code{"Y"}, latent responses \code{"Z"},
covariates \code{"X"} and their effects \code{"B"}, mandatory covariates
\code{"X0"} and their effects \code{"B0"}, latent indicator cocariates
\code{"Gamma"} and response graph \code{"G"}. The R code for generating the
simulated data is in file \code{"data-raw\simData.R"}.
}
\concept{datasets}
\keyword{datasets}
