% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.BayesMVP.R
\name{coef.BayesMVP}
\alias{coef.BayesMVP}
\title{coef method for class \code{BayesMVP}}
\usage{
\method{coef}{BayesMVP}(object, beta.type = "marginal", Pmax = 0, ...)
}
\arguments{
\item{object}{an object of class \code{BayesMVP}}

\item{beta.type}{type of output beta. Default is \code{marginal}, giving
marginal beta estimation. If \code{beta.type="conditional"}, it gives beta
estimation conditional on gamma=1.}

\item{Pmax}{If \code{Pmax=0.5} and \code{beta.type="conditional"}, it gives
median probability model betas. Default is 0.}

\item{...}{other arguments}
}
\value{
Estimated coefficients are from an object of class \code{BayesMVP}.
If the \code{BayesMVP} specified data standardization, the fitted values
are base based on standardized data.
}
\description{
Extract the posterior mean of the coefficients of a \code{BayesMVP} class object
}
\examples{
x <- 1

}
